Refactored GetEsVersionQuery. Refactored Requirementspage to SAS Pattern
